% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prolibr-package.R
\docType{package}
\name{prolibr-package}
\alias{prolibr}
\alias{prolibr-package}
\title{prolibr: Rstudio project library creation}
\description{
As of now package contains one function which creates project library
  directory inside Rstudio project directory and makes corresponding modification to
  project's .Rprofile file. This allows to install additional project specific 
  packages in project library by default with base command install.packages() 
  or devtools:: install_* function families when working under Rstudio project. 
  Thus project specific packages not polluting main library directories.
}
\author{
\strong{Maintainer}: Dmitrii Larionov \email{larionov.dimitrii.s@gmail.com}

}
\keyword{internal}
