% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{use_project_lib}
\alias{use_project_lib}
\title{Create Rstudio project library}
\usage{
use_project_lib(path = ".", lib = "library")
}
\arguments{
\item{path}{Character vector specifying path. May contain '.' for parent
directory reference.}

\item{lib}{Character vector specifying library directory name.}
}
\description{
Creates project library directory \code{lib} inside specified \code{path}.
}
\details{
If there is no .Rproj file inside specified \code{path} then error will be
thown.

If there is no .Rprofile file inside \code{path} then it will be created there
and activation command for project specific library will be written to it.
Otherwise activation command for project specific library will be appended to
the end of existing .Rprofile file.

If there is no \code{lib} directory inside \code{path} then directory will be created,
otherwise existing directory will be used as library with warning.
}
